pyproject.toml
[tool.poetry]
name = "tbk"
version = "0.0.0"
description = "奇遇淘客"
authors = ["dev <dev@qiyutech.tech>"]


[tool.poetry.dependencies]
python = "^3.9"

# 调度
apscheduler = ">=3.6,<4.0"

# web framework
django = ">=3.1,<3.2"

# for django database
psycopg2-binary = ">=2.8,<2.9"
# add MySQL if use MySQL database

# 线上环境展示
django-admin-env-notice = ">=0.2"

# 系统动态配置
django-constance = ">=2.8,<2.9"
# django-constance 依赖
django-picklefield = ">=3,<3.1"

# for django image field
pillow = ">=8.1,<9"

# for time zone
pytz = ">=2021,<2022"

# django production deployment
# should use uvicorn and use ASGI mod run python django
gunicorn = ">=20.0,<21"

# for log
django-structlog = ">=2.1,<2.2"

# dataclass <-> dict <-> json converter
dataclasses-json = ">=0.5,<0.6"

#########################################
# 当前 API 接口还没有稳定
# 不允许跨越小版本, 仅仅允许跨越 fix 版本
#########################################
# 大淘客接口
dtk-api = ">=0.7.1,<0.8"
# 折淘客接口
ztk-api = ">=0.3.2,<0.4"
#########################################

###################################################
# API 接口 依赖
fastapi = ">=0.63,<0.64"
# for fast API multipart
python-multipart = "<0.1"
# for deploy fastapi
uvicorn = ">=0.13,<0.14"
# for external API eg: 折淘客
aiohttp = ">=3.7,<4"
# for sync http request
requests = ">=2.24,<3"
###################################################


# for dot env files
django-dotenv = ">=1.4,<2.0"

django-qiyu-utils = ">=0.3.1,<0.4"
dj-qiyu-tpl = ">=0.3,<0.4"

[tool.poetry.dev-dependencies]
pytest = ">=6"
pytest-django = ">=4.1"
black = ">=20.8b1"
toml = ">=0.10,<1"
pylama = "*"

[build-system]
requires = ["poetry>=1"]
build-backend = "poetry.masonry.api"
